Publishing mockup-login/signup branch
I mostly have the page mockup done and am publishing the branch for it.
